pyproject.toml
[tool.poetry]
name = "event_maker"
version = "0.1.0"
description = ""
authors = ["Nazar <odddellarobea@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.5"
pytz = "^2022.1"
python-slugify = "^6.1.2"
Pillow = "^9.1.1"
argon2-cffi = "^21.3.0"
whitenoise = "^6.2.0"
redis = "^4.3.3"
celery = "^5.2.7"
django-celery-beat = "^2.3.0"
flower = "^1.1.0"
hiredis = "^2.0.0"
django = "^3.2.14"
django-environ = "^0.9.0"
django-model-utils = "^4.2.0"
django-redis = "^5.2.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.22.1"
psycopg2-binary = "^2.9.3"
django-rest-framework-social-oauth2 = "^1.1.0"
django-filter = "^22.1"

invoke = {version = "^1.7.1", optional = true}
rich = {version = "^12.5.1", optional = true}
ipdb = {version = "^0.13.9", optional = true}
Werkzeug = {version = "^2.0.3", optional = true}
watchfiles = {version = "^0.16.1", optional = true}
django-stubs = {version = "^1.9.0", optional = true}
pytest = {version = "^7.1.2", optional = true}
pytest-sugar = {version = "^0.9.4", optional = true}
pytest-django = {version = "^4.5.2", optional = true}
djangorestframework-stubs = {version = "^1.4.0", optional = true}
flake8 = {version = "^4.0.1", optional = true}
flake8-isort = {version = "^4.1.1", optional = true}
coverage = {version = "^6.4.1", optional = true}
black = {version = "^22.3.0", optional = true}
pre-commit = {version = "^2.20.0", optional = true}
factory-boy = {version = "^3.2.1", optional = true}
django-extensions = {version = "^3.1.5", optional = true}
django-debug-toolbar = {version = "^3.5.0", optional = true}
django-coverage-plugin = {version = "^2.0.3", optional = true}
uvicorn = {version = "^0.18.2", optional = true}

django-anymail = {version = "^8.6", optional = true}
gunicorn = {version = "^20.1.0", optional = true}

[tool.poetry.extras]
dev = [
    "invoke",
    "rich",
    "ipdb",
    "Werkzeug",
    "watchfiles",
    "django-stubs",
    "pytest",
    "pytest-sugar",
    "pytest-django",
    "djangorestframework-stubs",
    "flake8",
    "flake8-isort",
    "coverage",
    "black",
    "pre-commit",
    "factory-boy",
    "django-extensions",
    "django-debug-toolbar",
    "django-coverage-plugin",
    "uvicorn",
]
prod = [
    "django-anymail",
    "gunicorn",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
